Commented out element that causes an error
Commented out line 47 (account_vault_api_id). When set within PayForm it is being misidentified as an existing vault id rather than setting the vault API id for a newly vaulted account. This is a known issue with PayForm and is expected to be resolved in a future Paya Connect update.